KS-2-1 Added .toFloat() method. Renamed the file (it was 2-2, but should be 2-1)

+++ src/main/kotlin/sprintlesson_2/lesson2_task1.kt
@@ -7,7 +7,7 @@ fun main() {
     val student3 = 3
     val student4 = 5
 
-    val averageScore: Float = (student1 + student2 + student3 + student4) / 4f
+    val averageScore: Float = ((student1 + student2 + student3 + student4)).toFloat() / 4
 
     println( String.format("%.2f", averageScore) )
